cos/queue_ops: extract MQFQ V_min coordination into queue_ops/v_min.rs
#1034 Phase 1 — first split of the 4,531 prod LOC queue_ops monolith.
Convert queue_ops.rs to queue_ops/mod.rs (directory module) and
extract the V_min coordination cluster:

- publish_committed_queue_vtime
- compute_v_min_lag_threshold (private helper)
- cos_queue_v_min_consume_suspension
- cos_queue_v_min_continue

These 4 fns implement the per-queue virtual-time floor coordination
across workers participating in shared-exact queues — the suspension/
continuation handshake that prevents runaway flows from monopolizing
a shared-exact queue. Cohesive cluster, well-named, easy to lift
without disrupting the surrounding push/pop/drain code.

queue_ops/mod.rs production LOC (excluding mod tests):
  4,531 → 4,370 (still well over 2,000 threshold; this is just
  the first phase)
queue_ops/v_min.rs: 180 LOC.

Phase 2 candidates (not in this PR):
- Flow accounting (account_cos_queue_flow_*) ~90 prod LOC
- Drain orchestration (cos_queue_drain_all, cos_queue_restore_front,
  cos_queue_clear_orphan_snapshot_after_drop) ~90 prod LOC
- Push ops (cos_queue_push_*) ~200 prod LOC
- Pop ops (cos_queue_pop_*) ~165 prod LOC

build clean, 865 Rust tests pass. Cluster smoke required because
this touches the worker drain hot path.
